+++ src/docs/asciidoc/image-api.adoc
@@ -163,3 +163,60 @@ include::{snippets}/delete-image-success/response-fields.adoc[]
 - 삭제된 이미지는 일반 조회 API에서 제외됩니다
 - 본인이 업로드한 이미지만 삭제할 수 있습니다
 '''
+
+=== 이미지별 태그 조회 API
+==== GET /api/images/{imageId}/tags
+
+특정 이미지에 연결된 태그들을 조회합니다.
+
+===== 요청
+include::{snippets}/image-get-tags/http-request.adoc[]
+include::{snippets}/image-get-tags/path-parameters.adoc[]
+
+===== 응답
+include::{snippets}/image-get-tags/http-response.adoc[]
+include::{snippets}/image-get-tags/response-fields.adoc[]
+'''
+
+=== 이미지에 태그 추가 API
+==== POST /api/images/{imageId}/tags/{tagId}
+
+특정 이미지에 태그를 연결합니다.
+
+===== 요청
+include::{snippets}/image-add-tag/http-request.adoc[]
+include::{snippets}/image-add-tag/path-parameters.adoc[]
+
+===== 응답
+include::{snippets}/image-add-tag/http-response.adoc[]
+include::{snippets}/image-add-tag/response-fields.adoc[]
+'''
+
+=== 이미지에서 태그 제거 API
+==== DELETE /api/images/{imageId}/tags/{tagId}
+
+특정 이미지에서 태그 연결을 해제합니다.
+
+===== 요청
+include::{snippets}/image-remove-tag/http-request.adoc[]
+include::{snippets}/image-remove-tag/path-parameters.adoc[]
+
+===== 응답
+include::{snippets}/image-remove-tag/http-response.adoc[]
+include::{snippets}/image-remove-tag/response-fields.adoc[]
+'''
+
+=== 이미지 태그 수정 API
+==== PUT /api/images/{imageId}/tags/{tagId}
+
+특정 이미지에 연결된 태그를 다른 태그로 수정합니다.
+
+===== 요청
+include::{snippets}/image-update-tag/http-request.adoc[]
+include::{snippets}/image-update-tag/path-parameters.adoc[]
+include::{snippets}/image-update-tag/request-fields.adoc[]
+
+===== 응답
+include::{snippets}/image-update-tag/http-response.adoc[]
+include::{snippets}/image-update-tag/response-fields.adoc[]
+'''
