+++ src/docs/asciidoc/tag-api.adoc
@@ -2,7 +2,7 @@
 '''
 
 === 태그 생성 API
-==== POST /api/tag
+==== POST /api/tags
 
 새로운 태그를 생성합니다. 같은 이름의 기존 태그가 있으면 기존 태그를 반환합니다.
 
@@ -16,7 +16,7 @@ include::{snippets}/tag-create/response-fields.adoc[]
 '''
 
 === 모든 태그 조회 API
-==== GET /api/tag
+==== GET /api/tags
 
 시스템에 등록된 모든 태그를 조회합니다.
 
@@ -29,7 +29,7 @@ include::{snippets}/tag-get-all/response-fields.adoc[]
 '''
 
 === 시스템 태그 조회 API
-==== GET /api/tag/system
+==== GET /api/tags/system
 
 시스템에서 사전 정의된 태그들만 조회합니다.
 
@@ -41,65 +41,12 @@ include::{snippets}/tag-get-system/http-response.adoc[]
 include::{snippets}/tag-get-system/response-fields.adoc[]
 '''
 
-=== 이미지별 태그 조회 API
-==== GET /api/tag/image/{imageId}
 
-특정 이미지에 연결된 태그들을 조회합니다.
 
-===== 요청
-include::{snippets}/tag-get-by-image/http-request.adoc[]
-include::{snippets}/tag-get-by-image/path-parameters.adoc[]
 
-===== 응답
-include::{snippets}/tag-get-by-image/http-response.adoc[]
-include::{snippets}/tag-get-by-image/response-fields.adoc[]
-'''
-
-=== 이미지에 태그 추가 API
-==== POST /api/tag/image/{imageId}/tag/{tagId}
-
-특정 이미지에 태그를 연결합니다.
-
-===== 요청
-include::{snippets}/tag-add-to-image/http-request.adoc[]
-include::{snippets}/tag-add-to-image/path-parameters.adoc[]
-
-===== 응답
-include::{snippets}/tag-add-to-image/http-response.adoc[]
-include::{snippets}/tag-add-to-image/response-fields.adoc[]
-'''
-
-=== 이미지에서 태그 제거 API
-==== DELETE /api/tag/image/{imageId}/tag/{tagId}
-
-특정 이미지에서 태그 연결을 해제합니다.
-
-===== 요청
-include::{snippets}/tag-remove-from-image/http-request.adoc[]
-include::{snippets}/tag-remove-from-image/path-parameters.adoc[]
-
-===== 응답
-include::{snippets}/tag-remove-from-image/http-response.adoc[]
-include::{snippets}/tag-remove-from-image/response-fields.adoc[]
-'''
-
-=== 이미지 태그 수정 API
-==== PUT /api/tag/image/{imageId}/tag/{tagId}
-
-특정 이미지에 연결된 태그를 다른 태그로 수정합니다.
-
-===== 요청
-include::{snippets}/tag-update-image-tag/http-request.adoc[]
-include::{snippets}/tag-update-image-tag/path-parameters.adoc[]
-include::{snippets}/tag-update-image-tag/request-fields.adoc[]
-
-===== 응답
-include::{snippets}/tag-update-image-tag/http-response.adoc[]
-include::{snippets}/tag-update-image-tag/response-fields.adoc[]
-'''
 
 === 관리자 태그 삭제 API
-==== DELETE /api/tag/{tagId}
+==== DELETE /api/tags/admin/{tagId}
 
 관리자 권한으로 사용되지 않는 태그를 삭제합니다.
 - CUSTOM 타입 태그만 삭제 가능
@@ -115,7 +62,7 @@ include::{snippets}/tag-delete-by-admin/response-fields.adoc[]
 '''
 
 === 기본 태그 생성 API
-==== POST /api/tag/default
+==== POST /api/tags/default
 
 날짜/시간 정보를 기반으로 계절 및 시간대 태그를 자동 생성합니다.
